refactor comment endpoint
