+++ pages/api/comment.ts
@@ -4,6 +4,37 @@ import mongoose from "mongoose";
 import {commentModel, likeModel, notificationModel, updateModel, userModel} from "../../models/models";
 
 export default async function newCommentHandler(req: NextApiRequest, res: NextApiResponse) {
+    if (req.method === "GET") {
+        try {
+            if (Array.isArray(req.query.updateId)) return res.status(500).json({message: "Error: multiple update ids in query"});
+            const updateId: string = req.query.updateId;
+            if (!updateId) return res.status(500).json({message: "Cannot get profile, no update id given"});
+
+            await mongoose.connect(process.env.MONGODB_URL, {
+                useNewUrlParser: true,
+                useUnifiedTopology: true,
+                useFindAndModify: false,
+            });
+
+            const comments = await commentModel.aggregate([
+                {$match: {updateId: mongoose.Types.ObjectId(updateId)}},
+                {$lookup: {from: "users", foreignField: "_id", localField: "authorId", as: "authorArr"}},
+            ]);
+
+            const allMentionIds = comments.reduce((a, b) => {
+                const mentionStrings = b.body.match(/(?<=@\[).*?(?=\))/g);
+                const mentionIds = mentionStrings ? mentionStrings.map(d => d.split("](")[1]) : [];
+                return [...a, ...mentionIds];
+            }, []).filter((d, i, a) => a.findIndex(x => x ===d) === i);
+
+            const mentionedUsers = await userModel.find({_id: {$in: allMentionIds}});
+
+            return res.status(200).json({comments: comments, mentionedUsers: mentionedUsers});
+        } catch (e) {
+            return res.status(500).json({error: e});
+        }
+    }
+
     const session = await getSession({ req });
     if (!session) return res.status(403).json({message: "You must be signed in to post or delete an update."});
 
@@ -61,9 +92,9 @@ export default async function newCommentHandler(req: NextApiRequest, res: NextAp
                 }
             }
 
-            res.status(200).json({message: "success", data: returnComment});
+            return res.status(200).json({message: "success", data: returnComment});
         } catch (e) {
-            res.status(500).json({error: e});
+            return res.status(500).json({error: e});
         }
     } else if (req.method === "DELETE") {
         if (!req.body.commentId) return res.status(400).send("Missing commentId");
@@ -112,9 +143,9 @@ export default async function newCommentHandler(req: NextApiRequest, res: NextAp
 
             await commentModel.deleteMany({ _id: {$in: commentsToDeleteIds} });
             await likeModel.deleteMany({ _id: {$in: likesToDeleteIds} });
-            res.status(200).json({message: "success"});
+            return res.status(200).json({message: "success"});
         } catch (e) {
-            res.status(500).json({error: e});
+            return res.status(500).json({error: e});
         }
     } else {
         return res.status(405).send("Invalid method");
